Moving test cases into subdirectory
