+++ tests/testrunner.html
@@ -19,7 +19,7 @@
             'src/legion'
           ],
           'pixi':  'node_modules/pixi.js/bin/pixi',
-          'tests': 'tests'
+          'tests': 'tests/tests'
         }
     });
   </script>
