+++ tests/tests/graphics.js
